Define general 'post' prop type in table-row component
